refactor(mobile): move resolution engine out of MobileBridge into mobile_sync (#496 step 2)
